relocate components to proper components folder
